backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2.11
django-colorfield==0.11.0
django-extra-fields==3.0.2
django-filter==24.2
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-writable-nested==0.7.0
idna==3.6
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
webcolors==1.13
